+++ main/src/main/java/net/oneandone/stool/cli/ProjectCommand.java
@@ -17,8 +17,11 @@ package net.oneandone.stool.cli;
 
 import net.oneandone.stool.locking.Mode;
 import net.oneandone.stool.stage.Project;
+import net.oneandone.stool.util.Backstage;
 import net.oneandone.stool.util.Session;
 import net.oneandone.sushi.fs.file.FileNode;
+
+import java.io.IOException;
 
 public abstract class ProjectCommand extends SessionCommand {
     private final FileNode project;
@@ -30,8 +33,8 @@ public abstract class ProjectCommand extends SessionCommand {
 
     @Override
     public void doRun() throws Exception {
-        doRun(Project.load(project));
+        doRun(project);
     }
 
-    public abstract void doRun(Project project) throws Exception;
+    public abstract void doRun(FileNode project) throws Exception;
 }